Add files via upload
This shows how the model's performance changed when you over sample the data to balance it between the majority data points and minority data points. we have an improved performance from 69 percent without over-sample and 95 percent with over-sample. meaning i didn't have to use cross validation to improve the accuracy of the model.